refactored a bunch of paths after moving files
