+++ src/formik/inputs/tagSelector.tsx
@@ -3,13 +3,13 @@
 // ===================================================================
 import { useField } from 'formik';
 import React from 'react';
-import { Tag, TagState } from '../../../../../redux/tags/types';
+import { Tag, TagState } from '../../redux/tags/types';
 import { useSelector } from 'react-redux';
 
 import TextField from '@mui/material/TextField';
 import Autocomplete from '@mui/lab/Autocomplete';
-import TagChip from '../../../tags/tagChip';
-import { RootState } from '../../../../../redux/store';
+import TagChip from '../../components/tags/tagChip';
+import { RootState } from '../../redux/store';
 
 // ===================================================================
 //                            Component
