+++ src/pages/sign_up.tsx
@@ -5,10 +5,10 @@ import { useRouter } from 'next/router';
 
 import React from 'react';
 import { connect, ConnectedProps } from 'react-redux';
-import FormButton from '../components/app/shared/forms/building_blocks/formButton';
+import FormButton from '../formik/inputs/formButton';
 
-import TestComponent from '../components/debug';
-import PageLayout from '../components/pageLayout';
+import TestComponent from '../components/misc/debug';
+import PageLayout from '../components/layout/pageLayout';
 import { newUserCall } from '../api/user/call';
 import { loginAction } from '../redux/user/action';
